requirements.txt
numpy==1.19.4
Pillow==8.0.1
pywinusb==0.4.2
